[FLINK-11731] [State Backends] Make DefaultOperatorStateBackend follow the builder pattern
This closes #7899.
