+++ flink-runtime/src/main/java/org/apache/flink/runtime/state/DefaultOperatorStateBackend.java
@@ -18,6 +18,7 @@
 
 package org.apache.flink.runtime.state;
 
+import org.apache.commons.io.IOUtils;
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.state.BroadcastState;
@@ -26,20 +27,11 @@ import org.apache.flink.api.common.state.ListStateDescriptor;
 import org.apache.flink.api.common.state.MapStateDescriptor;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.common.typeutils.TypeSerializerSchemaCompatibility;
-import org.apache.flink.api.common.typeutils.UnloadableDummyTypeSerializer;
 import org.apache.flink.core.fs.CloseableRegistry;
-import org.apache.flink.core.fs.FSDataInputStream;
-import org.apache.flink.core.fs.FSDataOutputStream;
-import org.apache.flink.core.memory.DataInputView;
-import org.apache.flink.core.memory.DataInputViewStreamWrapper;
-import org.apache.flink.core.memory.DataOutputView;
-import org.apache.flink.core.memory.DataOutputViewStreamWrapper;
 import org.apache.flink.runtime.checkpoint.CheckpointOptions;
-import org.apache.flink.runtime.state.metainfo.StateMetaInfoSnapshot;
 import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.StateMigrationException;
 
-import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -47,13 +39,9 @@ import javax.annotation.Nonnull;
 
 import java.io.IOException;
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.FutureTask;
 import java.util.concurrent.RunnableFuture;
 
 /**
@@ -85,24 +73,14 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 	private final CloseableRegistry closeStreamOnCancelRegistry;
 
 	/**
-	 * Default serializer. Only used for the default operator state.
+	 * Default typeSerializer. Only used for the default operator state.
 	 */
-	private final JavaSerializer<Serializable> javaSerializer;
-
-	/**
-	 * The user code classloader.
-	 */
-	private final ClassLoader userClassloader;
+	private final JavaSerializer<Serializable> deprecatedDefaultJavaSerializer = new JavaSerializer<>();
 
 	/**
 	 * The execution configuration.
 	 */
 	private final ExecutionConfig executionConfig;
-
-	/**
-	 * Flag to de/activate asynchronous snapshots.
-	 */
-	private final boolean asynchronousSnapshots;
 
 	/**
 	 * Cache of already accessed states.
@@ -114,27 +92,27 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 	 *
 	 * @see <a href="https://issues.apache.org/jira/browse/FLINK-6849">FLINK-6849</a>
 	 */
-	private final HashMap<String, PartitionableListState<?>> accessedStatesByName;
+	private final Map<String, PartitionableListState<?>> accessedStatesByName;
 
 	private final Map<String, BackendWritableBroadcastState<?, ?>> accessedBroadcastStatesByName;
 
 	private final AbstractSnapshotStrategy<OperatorStateHandle> snapshotStrategy;
 
 	public DefaultOperatorStateBackend(
-		ClassLoader userClassLoader,
 		ExecutionConfig executionConfig,
-		boolean asynchronousSnapshots) {
-
-		this.closeStreamOnCancelRegistry = new CloseableRegistry();
-		this.userClassloader = Preconditions.checkNotNull(userClassLoader);
+		CloseableRegistry closeStreamOnCancelRegistry,
+		Map<String, PartitionableListState<?>> registeredOperatorStates,
+		Map<String, BackendWritableBroadcastState<?, ?>> registeredBroadcastStates,
+		Map<String, PartitionableListState<?>> accessedStatesByName,
+		Map<String, BackendWritableBroadcastState<?, ?>> accessedBroadcastStatesByName,
+		AbstractSnapshotStrategy<OperatorStateHandle> snapshotStrategy) {
+		this.closeStreamOnCancelRegistry = closeStreamOnCancelRegistry;
 		this.executionConfig = executionConfig;
-		this.javaSerializer = new JavaSerializer<>();
-		this.registeredOperatorStates = new HashMap<>();
-		this.registeredBroadcastStates = new HashMap<>();
-		this.asynchronousSnapshots = asynchronousSnapshots;
-		this.accessedStatesByName = new HashMap<>();
-		this.accessedBroadcastStatesByName = new HashMap<>();
-		this.snapshotStrategy = new DefaultOperatorStateBackendSnapshotStrategy();
+		this.registeredOperatorStates = registeredOperatorStates;
+		this.registeredBroadcastStates = registeredBroadcastStates;
+		this.accessedStatesByName = accessedStatesByName;
+		this.accessedBroadcastStatesByName = accessedBroadcastStatesByName;
+		this.snapshotStrategy = snapshotStrategy;
 	}
 
 	public ExecutionConfig getExecutionConfig() {
@@ -216,13 +194,13 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 			TypeSerializerSchemaCompatibility<K> keyCompatibility =
 				restoredBroadcastStateMetaInfo.updateKeySerializer(broadcastStateKeySerializer);
 			if (keyCompatibility.isIncompatible()) {
-				throw new StateMigrationException("The new key serializer for broadcast state must not be incompatible.");
+				throw new StateMigrationException("The new key typeSerializer for broadcast state must not be incompatible.");
 			}
 
 			TypeSerializerSchemaCompatibility<V> valueCompatibility =
 				restoredBroadcastStateMetaInfo.updateValueSerializer(broadcastStateValueSerializer);
 			if (valueCompatibility.isIncompatible()) {
-				throw new StateMigrationException("The new value serializer for broadcast state must not be incompatible.");
+				throw new StateMigrationException("The new value typeSerializer for broadcast state must not be incompatible.");
 			}
 
 			broadcastState.setStateMetaInfo(restoredBroadcastStateMetaInfo);
@@ -264,7 +242,7 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 	@Deprecated
 	@Override
 	public <T extends Serializable> ListState<T> getSerializableListState(String stateName) throws Exception {
-		return (ListState<T>) getListState(new ListStateDescriptor<>(stateName, javaSerializer));
+		return (ListState<T>) getListState(new ListStateDescriptor<>(stateName, deprecatedDefaultJavaSerializer));
 	}
 
 	// -------------------------------------------------------------------------------------------
@@ -272,120 +250,7 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 	// -------------------------------------------------------------------------------------------
 
 	public void restore(Collection<OperatorStateHandle> restoreSnapshots) throws Exception {
-
-		if (null == restoreSnapshots || restoreSnapshots.isEmpty()) {
-			return;
-		}
-
-		for (OperatorStateHandle stateHandle : restoreSnapshots) {
-
-			if (stateHandle == null) {
-				continue;
-			}
-
-			FSDataInputStream in = stateHandle.openInputStream();
-			closeStreamOnCancelRegistry.registerCloseable(in);
-
-			ClassLoader restoreClassLoader = Thread.currentThread().getContextClassLoader();
-
-			try {
-				Thread.currentThread().setContextClassLoader(userClassloader);
-				OperatorBackendSerializationProxy backendSerializationProxy =
-						new OperatorBackendSerializationProxy(userClassloader);
-
-				backendSerializationProxy.read(new DataInputViewStreamWrapper(in));
-
-				List<StateMetaInfoSnapshot> restoredOperatorMetaInfoSnapshots =
-						backendSerializationProxy.getOperatorStateMetaInfoSnapshots();
-
-				// Recreate all PartitionableListStates from the meta info
-				for (StateMetaInfoSnapshot restoredSnapshot : restoredOperatorMetaInfoSnapshots) {
-
-					final RegisteredOperatorStateBackendMetaInfo<?> restoredMetaInfo =
-						new RegisteredOperatorStateBackendMetaInfo<>(restoredSnapshot);
-
-					if (restoredMetaInfo.getPartitionStateSerializer() instanceof UnloadableDummyTypeSerializer) {
-
-						// must fail now if the previous serializer cannot be restored because there is no serializer
-						// capable of reading previous state
-						// TODO when eager state registration is in place, we can try to get a convert deserializer
-						// TODO from the newly registered serializer instead of simply failing here
-
-						throw new IOException("Unable to restore operator state [" + restoredSnapshot.getName() + "]." +
-							" The previous serializer of the operator state must be present; the serializer could" +
-							" have been removed from the classpath, or its implementation have changed and could" +
-							" not be loaded. This is a temporary restriction that will be fixed in future versions.");
-					}
-
-					PartitionableListState<?> listState = registeredOperatorStates.get(restoredSnapshot.getName());
-
-					if (null == listState) {
-						listState = new PartitionableListState<>(restoredMetaInfo);
-
-						registeredOperatorStates.put(listState.getStateMetaInfo().getName(), listState);
-					} else {
-						// TODO with eager state registration in place, check here for serializer migration strategies
-					}
-				}
-
-				// ... and then get back the broadcast state.
-				List<StateMetaInfoSnapshot> restoredBroadcastMetaInfoSnapshots =
-						backendSerializationProxy.getBroadcastStateMetaInfoSnapshots();
-
-				for (StateMetaInfoSnapshot restoredSnapshot : restoredBroadcastMetaInfoSnapshots) {
-
-					final RegisteredBroadcastStateBackendMetaInfo<?, ?> restoredMetaInfo =
-						new RegisteredBroadcastStateBackendMetaInfo<>(restoredSnapshot);
-
-					if (restoredMetaInfo.getKeySerializer() instanceof UnloadableDummyTypeSerializer ||
-						restoredMetaInfo.getValueSerializer() instanceof UnloadableDummyTypeSerializer) {
-
-						// must fail now if the previous serializer cannot be restored because there is no serializer
-						// capable of reading previous state
-						// TODO when eager state registration is in place, we can try to get a convert deserializer
-						// TODO from the newly registered serializer instead of simply failing here
-
-						throw new IOException("Unable to restore broadcast state [" + restoredSnapshot.getName() + "]." +
-								" The previous key and value serializers of the state must be present; the serializers could" +
-								" have been removed from the classpath, or their implementations have changed and could" +
-								" not be loaded. This is a temporary restriction that will be fixed in future versions.");
-					}
-
-					BackendWritableBroadcastState<? ,?> broadcastState = registeredBroadcastStates.get(restoredSnapshot.getName());
-
-					if (broadcastState == null) {
-						broadcastState = new HeapBroadcastState<>(restoredMetaInfo);
-
-						registeredBroadcastStates.put(broadcastState.getStateMetaInfo().getName(), broadcastState);
-					} else {
-						// TODO with eager state registration in place, check here for serializer migration strategies
-					}
-				}
-
-				// Restore all the states
-				for (Map.Entry<String, OperatorStateHandle.StateMetaInfo> nameToOffsets :
-						stateHandle.getStateNameToPartitionOffsets().entrySet()) {
-
-					final String stateName = nameToOffsets.getKey();
-
-					PartitionableListState<?> listStateForName = registeredOperatorStates.get(stateName);
-					if (listStateForName == null) {
-						BackendWritableBroadcastState<?, ?> broadcastStateForName = registeredBroadcastStates.get(stateName);
-						Preconditions.checkState(broadcastStateForName != null, "Found state without " +
-								"corresponding meta info: " + stateName);
-						deserializeBroadcastStateValues(broadcastStateForName, in, nameToOffsets.getValue());
-					} else {
-						deserializeOperatorStateValues(listStateForName, in, nameToOffsets.getValue());
-					}
-				}
-
-			} finally {
-				Thread.currentThread().setContextClassLoader(restoreClassLoader);
-				if (closeStreamOnCancelRegistry.unregisterCloseable(in)) {
-					IOUtils.closeQuietly(in);
-				}
-			}
-		}
+		// all restore work done in builder and nothing to do here
 	}
 
 	@Nonnull
@@ -403,112 +268,6 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 
 		snapshotStrategy.logSyncCompleted(streamFactory, syncStartTime);
 		return snapshotRunner;
-	}
-
-	/**
-	 * Implementation of operator list state.
-	 *
-	 * @param <S> the type of an operator state partition.
-	 */
-	static final class PartitionableListState<S> implements ListState<S> {
-
-		/**
-		 * Meta information of the state, including state name, assignment mode, and serializer
-		 */
-		private RegisteredOperatorStateBackendMetaInfo<S> stateMetaInfo;
-
-		/**
-		 * The internal list the holds the elements of the state
-		 */
-		private final ArrayList<S> internalList;
-
-		/**
-		 * A serializer that allows to perform deep copies of internalList
-		 */
-		private final ArrayListSerializer<S> internalListCopySerializer;
-
-		PartitionableListState(RegisteredOperatorStateBackendMetaInfo<S> stateMetaInfo) {
-			this(stateMetaInfo, new ArrayList<S>());
-		}
-
-		private PartitionableListState(
-				RegisteredOperatorStateBackendMetaInfo<S> stateMetaInfo,
-				ArrayList<S> internalList) {
-
-			this.stateMetaInfo = Preconditions.checkNotNull(stateMetaInfo);
-			this.internalList = Preconditions.checkNotNull(internalList);
-			this.internalListCopySerializer = new ArrayListSerializer<>(stateMetaInfo.getPartitionStateSerializer());
-		}
-
-		private PartitionableListState(PartitionableListState<S> toCopy) {
-
-			this(toCopy.stateMetaInfo.deepCopy(), toCopy.internalListCopySerializer.copy(toCopy.internalList));
-		}
-
-		public void setStateMetaInfo(RegisteredOperatorStateBackendMetaInfo<S> stateMetaInfo) {
-			this.stateMetaInfo = stateMetaInfo;
-		}
-
-		public RegisteredOperatorStateBackendMetaInfo<S> getStateMetaInfo() {
-			return stateMetaInfo;
-		}
-
-		public PartitionableListState<S> deepCopy() {
-			return new PartitionableListState<>(this);
-		}
-
-		@Override
-		public void clear() {
-			internalList.clear();
-		}
-
-		@Override
-		public Iterable<S> get() {
-			return internalList;
-		}
-
-		@Override
-		public void add(S value) {
-			Preconditions.checkNotNull(value, "You cannot add null to a ListState.");
-			internalList.add(value);
-		}
-
-		@Override
-		public String toString() {
-			return "PartitionableListState{" +
-					"stateMetaInfo=" + stateMetaInfo +
-					", internalList=" + internalList +
-					'}';
-		}
-
-		public long[] write(FSDataOutputStream out) throws IOException {
-
-			long[] partitionOffsets = new long[internalList.size()];
-
-			DataOutputView dov = new DataOutputViewStreamWrapper(out);
-
-			for (int i = 0; i < internalList.size(); ++i) {
-				S element = internalList.get(i);
-				partitionOffsets[i] = out.getPos();
-				getStateMetaInfo().getPartitionStateSerializer().serialize(element, dov);
-			}
-
-			return partitionOffsets;
-		}
-
-		@Override
-		public void update(List<S> values) {
-			internalList.clear();
-
-			addAll(values);
-		}
-
-		@Override
-		public void addAll(List<S> values) {
-			if (values != null && !values.isEmpty()) {
-				internalList.addAll(values);
-			}
-		}
 	}
 
 	private <S> ListState<S> getListState(
@@ -567,7 +326,7 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 			TypeSerializerSchemaCompatibility<S> stateCompatibility =
 				restoredPartitionableListStateMetaInfo.updatePartitionStateSerializer(newPartitionStateSerializer);
 			if (stateCompatibility.isIncompatible()) {
-				throw new StateMigrationException("The new state serializer for operator state must not be incompatible.");
+				throw new StateMigrationException("The new state typeSerializer for operator state must not be incompatible.");
 			}
 
 			partitionableListState.setStateMetaInfo(restoredPartitionableListStateMetaInfo);
@@ -575,47 +334,6 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 
 		accessedStatesByName.put(name, partitionableListState);
 		return partitionableListState;
-	}
-
-	private static <S> void deserializeOperatorStateValues(
-		PartitionableListState<S> stateListForName,
-		FSDataInputStream in,
-		OperatorStateHandle.StateMetaInfo metaInfo) throws IOException {
-
-		if (null != metaInfo) {
-			long[] offsets = metaInfo.getOffsets();
-			if (null != offsets) {
-				DataInputView div = new DataInputViewStreamWrapper(in);
-				TypeSerializer<S> serializer = stateListForName.getStateMetaInfo().getPartitionStateSerializer();
-				for (long offset : offsets) {
-					in.seek(offset);
-					stateListForName.add(serializer.deserialize(div));
-				}
-			}
-		}
-	}
-
-	private static <K, V> void deserializeBroadcastStateValues(
-			final BackendWritableBroadcastState<K, V> broadcastStateForName,
-			final FSDataInputStream in,
-			final OperatorStateHandle.StateMetaInfo metaInfo) throws Exception {
-
-		if (metaInfo != null) {
-			long[] offsets = metaInfo.getOffsets();
-			if (offsets != null) {
-
-				TypeSerializer<K> keySerializer = broadcastStateForName.getStateMetaInfo().getKeySerializer();
-				TypeSerializer<V> valueSerializer = broadcastStateForName.getStateMetaInfo().getValueSerializer();
-
-				in.seek(offsets[0]);
-
-				DataInputView div = new DataInputViewStreamWrapper(in);
-				int size = div.readInt();
-				for (int i = 0; i < size; i++) {
-					broadcastStateForName.put(keySerializer.deserialize(div), valueSerializer.deserialize(div));
-				}
-			}
-		}
 	}
 
 	private static void checkStateNameAndMode(
@@ -637,166 +355,4 @@ public class DefaultOperatorStateBackend implements OperatorStateBackend {
 				"registered with [" + expectedMode + "].");
 	}
 
-	/**
-	 * Snapshot strategy for this backend.
-	 */
-	private class DefaultOperatorStateBackendSnapshotStrategy extends AbstractSnapshotStrategy<OperatorStateHandle> {
-
-		protected DefaultOperatorStateBackendSnapshotStrategy() {
-			super("DefaultOperatorStateBackend snapshot");
-		}
-
-		@Nonnull
-		@Override
-		public RunnableFuture<SnapshotResult<OperatorStateHandle>> snapshot(
-			final long checkpointId,
-			final long timestamp,
-			@Nonnull final CheckpointStreamFactory streamFactory,
-			@Nonnull final CheckpointOptions checkpointOptions) throws IOException {
-
-			if (registeredOperatorStates.isEmpty() && registeredBroadcastStates.isEmpty()) {
-				return DoneFuture.of(SnapshotResult.empty());
-			}
-
-			final Map<String, PartitionableListState<?>> registeredOperatorStatesDeepCopies =
-				new HashMap<>(registeredOperatorStates.size());
-			final Map<String, BackendWritableBroadcastState<?, ?>> registeredBroadcastStatesDeepCopies =
-				new HashMap<>(registeredBroadcastStates.size());
-
-			ClassLoader snapshotClassLoader = Thread.currentThread().getContextClassLoader();
-			Thread.currentThread().setContextClassLoader(userClassloader);
-			try {
-				// eagerly create deep copies of the list and the broadcast states (if any)
-				// in the synchronous phase, so that we can use them in the async writing.
-
-				if (!registeredOperatorStates.isEmpty()) {
-					for (Map.Entry<String, PartitionableListState<?>> entry : registeredOperatorStates.entrySet()) {
-						PartitionableListState<?> listState = entry.getValue();
-						if (null != listState) {
-							listState = listState.deepCopy();
-						}
-						registeredOperatorStatesDeepCopies.put(entry.getKey(), listState);
-					}
-				}
-
-				if (!registeredBroadcastStates.isEmpty()) {
-					for (Map.Entry<String, BackendWritableBroadcastState<?, ?>> entry : registeredBroadcastStates.entrySet()) {
-						BackendWritableBroadcastState<?, ?> broadcastState = entry.getValue();
-						if (null != broadcastState) {
-							broadcastState = broadcastState.deepCopy();
-						}
-						registeredBroadcastStatesDeepCopies.put(entry.getKey(), broadcastState);
-					}
-				}
-			} finally {
-				Thread.currentThread().setContextClassLoader(snapshotClassLoader);
-			}
-
-			AsyncSnapshotCallable<SnapshotResult<OperatorStateHandle>> snapshotCallable =
-				new AsyncSnapshotCallable<SnapshotResult<OperatorStateHandle>>() {
-
-					@Override
-					protected SnapshotResult<OperatorStateHandle> callInternal() throws Exception {
-
-						CheckpointStreamFactory.CheckpointStateOutputStream localOut =
-							streamFactory.createCheckpointStateOutputStream(CheckpointedStateScope.EXCLUSIVE);
-						snapshotCloseableRegistry.registerCloseable(localOut);
-
-						// get the registered operator state infos ...
-						List<StateMetaInfoSnapshot> operatorMetaInfoSnapshots =
-							new ArrayList<>(registeredOperatorStatesDeepCopies.size());
-
-						for (Map.Entry<String, PartitionableListState<?>> entry :
-							registeredOperatorStatesDeepCopies.entrySet()) {
-							operatorMetaInfoSnapshots.add(entry.getValue().getStateMetaInfo().snapshot());
-						}
-
-						// ... get the registered broadcast operator state infos ...
-						List<StateMetaInfoSnapshot> broadcastMetaInfoSnapshots =
-							new ArrayList<>(registeredBroadcastStatesDeepCopies.size());
-
-						for (Map.Entry<String, BackendWritableBroadcastState<?, ?>> entry :
-							registeredBroadcastStatesDeepCopies.entrySet()) {
-							broadcastMetaInfoSnapshots.add(entry.getValue().getStateMetaInfo().snapshot());
-						}
-
-						// ... write them all in the checkpoint stream ...
-						DataOutputView dov = new DataOutputViewStreamWrapper(localOut);
-
-						OperatorBackendSerializationProxy backendSerializationProxy =
-							new OperatorBackendSerializationProxy(operatorMetaInfoSnapshots, broadcastMetaInfoSnapshots);
-
-						backendSerializationProxy.write(dov);
-
-						// ... and then go for the states ...
-
-						// we put BOTH normal and broadcast state metadata here
-						int initialMapCapacity =
-							registeredOperatorStatesDeepCopies.size() + registeredBroadcastStatesDeepCopies.size();
-						final Map<String, OperatorStateHandle.StateMetaInfo> writtenStatesMetaData =
-							new HashMap<>(initialMapCapacity);
-
-						for (Map.Entry<String, PartitionableListState<?>> entry :
-							registeredOperatorStatesDeepCopies.entrySet()) {
-
-							PartitionableListState<?> value = entry.getValue();
-							long[] partitionOffsets = value.write(localOut);
-							OperatorStateHandle.Mode mode = value.getStateMetaInfo().getAssignmentMode();
-							writtenStatesMetaData.put(
-								entry.getKey(),
-								new OperatorStateHandle.StateMetaInfo(partitionOffsets, mode));
-						}
-
-						// ... and the broadcast states themselves ...
-						for (Map.Entry<String, BackendWritableBroadcastState<?, ?>> entry :
-							registeredBroadcastStatesDeepCopies.entrySet()) {
-
-							BackendWritableBroadcastState<?, ?> value = entry.getValue();
-							long[] partitionOffsets = {value.write(localOut)};
-							OperatorStateHandle.Mode mode = value.getStateMetaInfo().getAssignmentMode();
-							writtenStatesMetaData.put(
-								entry.getKey(),
-								new OperatorStateHandle.StateMetaInfo(partitionOffsets, mode));
-						}
-
-						// ... and, finally, create the state handle.
-						OperatorStateHandle retValue = null;
-
-						if (snapshotCloseableRegistry.unregisterCloseable(localOut)) {
-
-							StreamStateHandle stateHandle = localOut.closeAndGetHandle();
-
-							if (stateHandle != null) {
-								retValue = new OperatorStreamStateHandle(writtenStatesMetaData, stateHandle);
-							}
-
-							return SnapshotResult.of(retValue);
-						} else {
-							throw new IOException("Stream was already unregistered.");
-						}
-					}
-
-					@Override
-					protected void cleanupProvidedResources() {
-						// nothing to do
-					}
-
-					@Override
-					protected void logAsyncSnapshotComplete(long startTime) {
-						if (asynchronousSnapshots) {
-							logAsyncCompleted(streamFactory, startTime);
-						}
-					}
-				};
-
-			final FutureTask<SnapshotResult<OperatorStateHandle>> task =
-				snapshotCallable.toAsyncSnapshotFutureTask(closeStreamOnCancelRegistry);
-
-			if (!asynchronousSnapshots) {
-				task.run();
-			}
-
-			return task;
-		}
-	}
 }
