+++ flink-runtime/src/test/java/org/apache/flink/runtime/state/StateBackendMigrationTestBase.java
@@ -28,6 +28,7 @@ import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.common.typeutils.TypeSerializerSchemaCompatibility;
 import org.apache.flink.api.common.typeutils.TypeSerializerSnapshot;
 import org.apache.flink.api.common.typeutils.base.IntSerializer;
+import org.apache.flink.core.fs.CloseableRegistry;
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 import org.apache.flink.runtime.checkpoint.CheckpointOptions;
@@ -46,6 +47,7 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 import java.io.IOException;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
@@ -1059,12 +1061,17 @@ public abstract class StateBackendMigrationTestBase<B extends AbstractStateBacke
 	// -------------------------------------------------------------------------------
 
 	private OperatorStateBackend createOperatorStateBackend() throws Exception {
-		return getStateBackend().createOperatorStateBackend(new DummyEnvironment(), "test_op");
+		return getStateBackend().createOperatorStateBackend(
+			new DummyEnvironment(), "test_op", Collections.emptyList(), new CloseableRegistry());
+	}
+
+	private OperatorStateBackend createOperatorStateBackend(Collection<OperatorStateHandle> state) throws Exception {
+		return getStateBackend().createOperatorStateBackend(
+			new DummyEnvironment(), "test_op", state, new CloseableRegistry());
 	}
 
 	private OperatorStateBackend restoreOperatorStateBackend(OperatorStateHandle state) throws Exception {
-		OperatorStateBackend operatorStateBackend = createOperatorStateBackend();
-		operatorStateBackend.restore(StateObjectCollection.singleton(state));
+		OperatorStateBackend operatorStateBackend = createOperatorStateBackend(StateObjectCollection.singleton(state));
 		return operatorStateBackend;
 	}
